plugin.cpp: Restore debugging level. gui/gui.cpp: Don't convert range from int to float. gui/gui.h: Add some documentation for FD watching API. gui/gtksup.h: Remove defunct documentation.

+++ gui/gui.cpp
@@ -1081,7 +1081,7 @@ void
 Gui::setInvalidatedRegions(const InvalidatedRanges& ranges)
 {
     // fallback to single regions
-    geometry::Range2d<float> full = ranges.getFullArea();
+    geometry::Range2d<int> full = ranges.getFullArea();
     
     SWFRect bounds;
